Add Score abstract class; add Score2048 class
