+++ Game2048/Game/Controllers/Game2048.cs
@@ -6,12 +6,12 @@ public class Game2048 : Game
 {
     public Grid? grid;
 
-    public new Score? score;
+    public Score? score;
 
     public override void Start()
     {
         grid = new Grid(4, 4);
-        score = new Score();
+        score = new Score2048();
     }
 
     public Game2048()
