+++ Game2048/Game/Models/Abstracts/Game.cs
@@ -2,8 +2,6 @@
 
 public abstract class Game
 {
-    public Score score = new Score();
-
     public abstract void Start();
 
 }
