+++ Game2048/Game/Models/Abstracts/Score.cs
@@ -1,0 +1,7 @@
+﻿namespace Models;
+
+public abstract class Score
+{
+    public int Value { get; set; }
+    public abstract void UpdateScore(int points);
+}
