--- Game2048/Game/Models/Score.cs
@@ -1,15 +1,0 @@
-﻿namespace Models;
-
-public class Score
-{
-    public int Value = 0;
-
-    public void AddScore(int points)
-    {
-        Value += points;
-        if (Value < 0)
-        {
-            Value = 0;
-        }
-    }
-}
