+++ Game2048/Game/Models/Score2048.cs
@@ -1,0 +1,20 @@
+﻿namespace Models;
+
+public class Score2048 : Score
+{
+    public new int Value;
+
+    public override void UpdateScore(int points)
+    {
+        Value = points;
+        if (Value < 0)
+        {
+            Value = 0;
+        }
+    }
+
+    public Score2048()
+    {
+        Value = 0;
+    }
+}
