+++ Game2048/Game/Pages/Index.razor.css
